+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is inspired by *Keep a Changelog* while remaining focused on practica
 ## Changed
 - sgnd-install.sh to never overwrite metadata of existing folders in teh target tree
 - prepare-release.sh to set ownership of repository folders to root:root
+- added toggle button to sgnd-console to switch between root and non-root
 
 # Version 1.8 (Build 2620413) 2026-07-23